+++ inc/utility-functions.php
@@ -625,12 +625,13 @@ function wolf_core_sanitize_heading( $string ) {
 
 	$post_title = wolf_core_get_post_title();
 	$site_title = get_bloginfo( 'name' );
-	// $post_title = get_the_title();
+	$post_excerpt = get_the_excerpt();
 
 	$short_tags = array(
 		'{{post_title}}' => $post_title,
 		'{{page_title}}' => $post_title,
 		'{{site_title}}' => $site_title,
+		'{{post_excerpt}}' => $post_excerpt,
 	);
 
 	foreach ( $short_tags as $key => $value ) {