+++ plugins/elementor/controls/heading.php
@@ -22,11 +22,13 @@ add_action(
 	function( $content ) {
 		$post_title = wolf_core_get_post_title();
 		$site_title = get_bloginfo( 'name' );
+		$post_excerpt = get_the_excerpt();
 
 		$short_tags = array(
 			'{{post_title}}' => $post_title,
 			'{{page_title}}' => $post_title,
 			'{{site_title}}' => $site_title,
+			'{{post_excerpt}}' => $post_excerpt,
 		);
 
 		foreach ( $short_tags as $key => $value ) {